requirements.txt
Django>=1.4
coverage==3.7.1
django-coverage==1.2.4
djangorestframework > 3.6.0
django-filter
django-crispy-forms
#wsgiref==0.1.2
